+++ javascript/node/selenium-webdriver/io/index.js
@@ -208,20 +208,21 @@ exports.tmpFile = function(opt_options) {
  *     not be found.
  */
 exports.findInPath = function(file, opt_checkCwd) {
+  let dirs = [];
   if (opt_checkCwd) {
-    var tmp = path.join(process.cwd(), file);
-    if (fs.existsSync(tmp)) {
-      return tmp;
-    }
+    dirs.push(process.cwd());
   }
+  dirs.push.apply(dirs, process.env['PATH'].split(path.delimiter));
 
-  var dirs = process.env['PATH'].split(path.delimiter);
-  var found = null;
-  dirs.forEach(function(dir) {
-    var tmp = path.join(dir, file);
-    if (!found && fs.existsSync(tmp)) {
-      found = tmp;
+  let foundInDir = dirs.find(dir => {
+    let tmp = path.join(dir, file);
+    try {
+      let stats = fs.statSync(tmp);
+      return stats.isFile() && !stats.isDirectory();
+    } catch (ex) {
+      return false;
     }
   });
-  return found;
+
+  return foundInDir ? path.join(foundInDir, file) : null;
 };